architecture-function/requirements.txt
google-cloud-compute==1.19.1
redis==5.0.1
functions-framework==3.5.0
flask==3.0.0
google-auth==2.28.2